+++ changelog.rst
@@ -18,6 +18,7 @@ next release
 * Fix interact on python 3 in Flow.
 * Add python bytecode manipulation functions.
 * Added shellcode generator for X86/X86_64, ARM (+Thumb) and AArch64.
+* Use keystone engine as assembler engine by default.
 
 0.7.2 (2016-03-11)
 ==================